Refactor App as function component and useContext
